+++ ui/main/game/galactic_war/cards/gwaio_start_nomad.js
@@ -133,6 +133,20 @@ define([
               }
             );
           });
+          var orbitalStructures = [
+            "/pa/units/orbital/ion_defense/ion_defense.json",
+            "/pa/units/orbital/deep_space_radar/deep_space_radar.json",
+            "/pa/units/orbital/defense_satellite/defense_satellite.json",
+            "/pa/units/orbital/mining_platform/mining_platform.json",
+          ];
+          orbitalStructures.forEach(function (unit) {
+            mods.push({
+              file: unit,
+              path: "navigation.type",
+              op: "replace",
+              value: "orbital",
+            });
+          });
           inventory.addMods(mods);
         } else {
           inventory.maxCards(inventory.maxCards() + 1);